@@ -9,41 +9,41 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "postinstall": "npm --prefix ./data install ./data && npm --prefix ./functions install ./functions",
+    "postinstall": "npm --prefix ./data install ./data",
     "deploy": "npm run build && firebase deploy --project default",
     "deploy:travis": "npm run build && firebase deploy --project default --non-interactive --token $FIREBASE_TOKEN"
   },
   "private": false,
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.3",
-    "@angular/animations": "^6.0.3",
-    "@angular/common": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
-    "@angular/core": "^6.0.3",
-    "@angular/forms": "^6.0.3",
-    "@angular/http": "^6.0.3",
-    "@angular/platform-browser": "^6.0.3",
-    "@angular/platform-browser-dynamic": "^6.0.3",
-    "@angular/router": "^6.0.3",
-    "core-js": "^2.5.4",
-    "firebase": "^5.0.4",
-    "geofire": "^4.1.2",
-    "geokit": "0.0.3",
+    "@agm/core": "^1.0.0-beta.5",
+    "@angular/animations": "^6.1.9",
+    "@angular/common": "^6.1.9",
+    "@angular/compiler": "^6.1.9",
+    "@angular/core": "^6.1.9",
+    "@angular/forms": "^6.1.9",
+    "@angular/http": "^6.1.9",
+    "@angular/platform-browser": "^6.1.9",
+    "@angular/platform-browser-dynamic": "^6.1.9",
+    "@angular/router": "^6.1.9",
+    "core-js": "^2.5.7",
+    "firebase": "^5.5.3",
+    "geofirestore": "^2.2.1",
+    "geokit": "0.1.3",
     "normalize.css": "^8.0.0",
-    "rxjs": "^6.2.0",
-    "rxjs-compat": "^6.2.0",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.3",
-    "@angular/cli": "~6.0.3",
-    "@angular/compiler-cli": "^6.0.3",
-    "@angular/language-service": "^6.0.3",
+    "@angular-devkit/build-angular": "~0.7.2",
+    "@angular/cli": "~6.1.2",
+    "@angular/compiler-cli": "^6.1.3",
+    "@angular/language-service": "^6.1.3",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
     "codelyzer": "~4.2.1",
-    "firebase-tools": "^3.18.5",
+    "firebase-tools": "^4.0.3",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~1.7.1",
