@@ -1,10 +1,9 @@
 import { Injectable } from '@angular/core';
-import { GeoFire } from '../geofire/index.esm';
-import { Observable, BehaviorSubject } from 'rxjs';
-import { LatLngLiteral } from '@agm/core';
-import { Geokit } from 'geokit';
 import * as firebase from 'firebase/app';
-import 'firebase/database';
+import { GeoFirestore, GeoFirestoreQuery } from 'geofirestore';
+import { Geokit, LatLngLiteral } from 'geokit';
+import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+import { Observable } from 'rxjs/Observable';
 
 import { LocationService } from './location.service';
 
@@ -19,47 +18,56 @@ function quicksort(c: any[]): any {
 
 @Injectable()
 export class RestaurantsService {
-  private _geoFire: GeoFire;
-  private _previousCoords: LatLngLiteral = { lat: 0, lng: 0 };
-  private _restaurant: BehaviorSubject<any> = new BehaviorSubject<any>({});
+  private _geoFirestore: GeoFirestore;
+  private _previousCoords: firebase.firestore.GeoPoint = new firebase.firestore.GeoPoint(0, 0);
   private _restaurants: BehaviorSubject<any[]> = new BehaviorSubject<any[]>([]);
+  private _query: GeoFirestoreQuery;
 
   constructor(private _ls: LocationService) {
-    this._geoFire = new GeoFire(firebase.database().ref('geofire/restaurants'));
-    this._ls.mapCenter.subscribe((coords: LatLngLiteral) => {
-      if (Geokit.distance(coords, this._previousCoords) > 0.5) {
+    this._geoFirestore = new GeoFirestore(firebase.firestore().collection('restaurants'));
+    this._ls.mapCenter.subscribe((coords: firebase.firestore.GeoPoint) => {
+      if (Geokit.distance(this._geopoint2Literal(coords), this._geopoint2Literal(this._previousCoords)) > 0.5) {
         this._previousCoords = coords;
         this._geoFetch(coords);
       }
     });
   }
 
-  get restaurant(): Observable<any> {
-    return this._restaurant.asObservable();
-  }
-
   get restaurants(): Observable<any[]> {
     return this._restaurants.asObservable();
   }
 
-  private _geoFetch(coords: LatLngLiteral): void {
+  private _geoFetch(center: firebase.firestore.GeoPoint): void {
     const max = 100;
-    this._geoFire.query({
-      center: [coords.lat, coords.lng],
-      radius: 0.5
-    }).on('key_entered', (key: string, result: any[]) => {
+    if (this._query) {
+      this._query.cancel();
+      this._query = null;
+    }
+    this._query = this._geoFirestore.query({
+      center,
+      radius: 1
+    });
+
+    this._query.on('key_entered', ($key: string, result: any) => {
       let places: any[] = [...this._restaurants.value];
-      const newPlace: any = { 'id': key, 'coords': { 'lat': result[0], 'lng': result[1] } };
-      if (places.find((place: any) => place.id === newPlace.id)) { return; }
-      places.push(newPlace);
-      places.map((place: any) => place.distance = Geokit.distance(coords, place.coords, 'miles'));
+      result.$key = $key;
+      if (places.find((place: any) => place.id === result.$key)) { return; }
+      places.push(result);
+      places.map((place: any) => place.distance = Geokit.distance(
+        this._geopoint2Literal(center),
+        this._geopoint2Literal(place.coordinates),
+        'miles')
+      );
       places = quicksort(places);
       if (places.length > max) { places = places.slice(0, max); }
       this._restaurants.next(places);
     });
   }
 
-  public updateRestaurant(id: string): void {
-    firebase.database().ref('restaurants').child(id).on('value', (snapshot: any) => this._restaurant.next(snapshot.val()));
+  private _geopoint2Literal(coordinates: firebase.firestore.GeoPoint): LatLngLiteral {
+    return {
+      lat: coordinates.latitude,
+      lng: coordinates.longitude
+    };
   }
 }
